Move code using Any type to separate submodule
